+++ config/api-client.php
@@ -18,6 +18,8 @@ return [
         ConfigKey::API_CLIENT => [
             ConfigKey::ENDPOINTS => [
                 Endpoint\Auth\AuthEndpoint::class,
+                Endpoint\Export\ExportCreateEndpoint::class,
+                Endpoint\Export\ExportStatusEndpoint::class,
                 Endpoint\Generic\GenericDetailsEndpoint::class,
                 Endpoint\Generic\GenericIconEndpoint::class,
                 Endpoint\Item\ItemIngredientEndpoint::class,