+++ src/Endpoint/Export/ExportCreateEndpoint.php
@@ -1,0 +1,54 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FactorioItemBrowser\Api\Client\Endpoint\Export;
+
+use FactorioItemBrowser\Api\Client\Endpoint\EndpointInterface;
+use FactorioItemBrowser\Api\Client\Request\Export\ExportCreateRequest;
+use FactorioItemBrowser\Api\Client\Response\Export\ExportCreateResponse;
+
+/**
+ * The endpoint of the export create request.
+ *
+ * @author BluePsyduck <bluepsyduck@gmx.com>
+ * @license http://opensource.org/licenses/GPL-3.0 GPL v3
+ */
+class ExportCreateEndpoint implements EndpointInterface
+{
+    /**
+     * Returns the request class supported by the endpoint.
+     * @return string
+     */
+    public function getSupportedRequestClass(): string
+    {
+        return ExportCreateRequest::class;
+    }
+
+    /**
+     * Returns whether or not this endpoint requires an authorization token.
+     * @return bool
+     */
+    public function requiresAuthorizationToken(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Returns the request path of the endpoint.
+     * @return string
+     */
+    public function getRequestPath(): string
+    {
+        return 'export/create';
+    }
+
+    /**
+     * Creates the response of the endpoint.
+     * @return string
+     */
+    public function getResponseClass(): string
+    {
+        return ExportCreateResponse::class;
+    }
+}