+++ src/Request/Export/ExportCreateRequest.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FactorioItemBrowser\Api\Client\Request\Export;
+
+use FactorioItemBrowser\Api\Client\Request\RequestInterface;
+
+/**
+ * The request to create a new export.
+ *
+ * @author BluePsyduck <bluepsyduck@gmx.com>
+ * @license http://opensource.org/licenses/GPL-3.0 GPL v3
+ */
+class ExportCreateRequest implements RequestInterface
+{
+}