+++ test/serializer/Request/Export/ExportStatusRequestTest.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FactorioItemBrowserTestSerializer\Api\Client\Request\Export;
+
+use FactorioItemBrowser\Api\Client\Request\Export\ExportStatusRequest;
+use FactorioItemBrowserTestAsset\Api\Client\SerializerTestCase;
+
+/**
+ * The PHPUnit test of serializing the ExportStatusRequest class.
+ *
+ * @author BluePsyduck <bluepsyduck@gmx.com>
+ * @license http://opensource.org/licenses/GPL-3.0 GPL v3
+ * @coversNothing
+ */
+class ExportStatusRequestTest extends SerializerTestCase
+{
+    /**
+     * Returns the object to be serialized or deserialized.
+     * @return object
+     */
+    protected function getObject(): object
+    {
+        $result = new ExportStatusRequest();
+        return $result;
+    }
+
+    /**
+     * Returns the serialized data.
+     * @return array
+     */
+    protected function getData(): array
+    {
+        return [
+        ];
+    }
+}